+++ django/applications/catmaid/urls.py
@@ -145,6 +145,7 @@ urlpatterns += [
 
 # Statistics
 urlpatterns += [
+    url(r'^(?P<project_id>\d+)/stats/cable-length$', stats.stats_cable_length),
     url(r'^(?P<project_id>\d+)/stats/nodecount$', stats.stats_nodecount),
     url(r'^(?P<project_id>\d+)/stats/editor$', stats.stats_editor),
     url(r'^(?P<project_id>\d+)/stats/summary$', stats.stats_summary),